refactor(struct): extract EPUB readable fallback
